+++ packages/playwright/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased (minor)
 
 - Update dependencies (including `typescript`)
+- Support [Serial Mode](https://playwright.dev/docs/test-parallel#serial-mode) by marking as failed the whole file for Retry-only-Failures
 
 ## 0.2.0